Date extension refactoring
- Simplify theory data in tests
- Rename tests to follow convention elsewhere
- Use the Type directory instead of TypeExtensions
- Rename Date.cs to DatePartial.cs to follow convention elsewhere
